+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 **Fixed bugs:**
 
+- Cachix caches aren't added to the configuration [\#208](https://github.com/divnix/devos/issues/208)
 - iso: systemd service startup [\#194](https://github.com/divnix/devos/issues/194)
 - Help adding easy-hls-nix to devos [\#174](https://github.com/divnix/devos/issues/174)
 - `flk update` fails because of obsolete flag [\#159](https://github.com/divnix/devos/issues/159)
@@ -40,6 +41,7 @@
 
 **Merged pull requests:**
 
+- hosts: add cachix [\#210](https://github.com/divnix/devos/pull/210)
 - iso: avoid systemd service startup [\#202](https://github.com/divnix/devos/pull/202)
 - hosts/devosSystem: pass modules as attrset [\#198](https://github.com/divnix/devos/pull/198)
 - doc: enact bootstrapping section [\#196](https://github.com/divnix/devos/pull/196)
@@ -47,6 +49,7 @@
 - iso: copy input closourse into iso to avoide re-download [\#191](https://github.com/divnix/devos/pull/191)
 - add hosts module arg [\#164](https://github.com/divnix/devos/pull/164)
 - Multi-arch support for hosts\(nixosConfigurations\) [\#161](https://github.com/divnix/devos/pull/161)
+- Improve Home Manager support: profiles/suites, modules, extern, flake outputs [\#156](https://github.com/divnix/devos/pull/156)
 
 ## [v0.8.0](https://github.com/divnix/devos/tree/v0.8.0) (2021-03-02)